Bugfix: convert LSP client IDs to strings before storing in the context
Thanks @sam4llis for helping me to get to the bottom of this bug!

It appears that Lua tables don't support integer keys where the values
don't start from 1, ascending. The result was that, in certain
situations where you have multiple LSP clients attached to a buffer, the
plugin would break because the context variable would not match this
rule.

The solution is to convert the LSP client IDs to strings before using
them as keys in the context table.